Updated the Pid->sequence validation to account for the fact that a single record can have multiple LogEntries and multiple pids on followers in the replication case.
We validate that there are no duplicates on the followers as well. But
if there are, it is a fatal exception, because there is nothing we can
do to recover it.

Further, if we recevie a duplicate on the the leader for any entry
except the last appended entry, it is also a fatal exception. It
indicates a bad client.