Couple of changes:
  1. Refactored the `RecordBatch.closeWithSequence` to be
`RecordBatch.closeWithTransactionState`. This sets the pid, epoch, and
sequence. We also throw an IllegalStateException if the batch was
already closed and the incoming pid, epoch, or sequence is different
from the existing ones.

  2. Added more test cases to SenderTest. It should now cover all the
new state changes introduced by the idempotent producer. With the unit
tests on the broker, plus the newly beefed up integration test, over
coverage for the idempotent producer is getting pretty good!